requirements.txt
tensorflow>=2.12.0
streamlit>=1.28.0
Pillow>=9.0.0
numpy>=1.23.0
